# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AccountExport": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/Member"
          },
          "settings": {
            "$ref": "#/components/schemas/UserSettings"
          },
          "sims": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SimFull"
            },
            "title": "Sims",
            "type": "array"
          }
        },
        "required": [
          "profile",
          "settings"
        ],
        "title": "AccountExport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Invite": {
        "properties": {
          "claimedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimedat"
          },
          "claimedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimedby"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "expired",
              "claimed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdBy",
          "createdAt",
          "expiresAt",
          "claimedBy",
          "claimedAt",
          "status"
        ],
        "title": "Invite",
        "type": "object"
      },
      "InviteCreate": {
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^\\s*[^@\\s]+@[^@\\s]+\\.[^@\\s]+\\s*$",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "default": "member",
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "InviteCreate",
        "type": "object"
      },
...
      "Me": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "uid": {
            "title": "Uid",
            "type": "string"
          }
        },
        "required": [
          "uid",
          "email",
          "displayName",
          "role"
        ],
        "title": "Me",
        "type": "object"
      },
      "Member": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "lastSeenAt": {
            "format": "date-time",
            "title": "Lastseenat",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "uid": {
            "title": "Uid",
            "type": "string"
          }
        },
        "required": [
          "uid",
          "email",
          "displayName",
          "role",
          "createdAt",
          "lastSeenAt"
        ],
        "title": "Member",
        "type": "object"
      },
...
      "SettingsPatch": {
        "additionalProperties": false,
        "description": "Only the fields sent change. Send with If-Match: the ETag from GET /me/settings.",
        "properties": {
          "alerts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerts"
          },
          "awakeEnd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awakeend"
          },
          "awakeStart": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awakestart"
          },
          "briefEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Briefenabled"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "draftStrategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draftstrategy"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "timeZone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "SettingsPatch",
        "type": "object"
      },
      "SimFull": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "hasDetail": {
            "description": "False once an old practice run is trimmed to its summary",
            "title": "Hasdetail",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "practice",
              "league"
            ],
            "title": "Kind",
            "type": "string"
          },
          "pinned": {
            "title": "Pinned",
            "type": "boolean"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "season",
          "title",
          "summary",
          "pinned",
          "hasDetail",
          "createdAt",
          "updatedAt",
          "version",
          "detail"
        ],
        "title": "SimFull",
        "type": "object"
      },
...
      "TelegramLink": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          }
        },
        "required": [
          "code",
          "expiresAt",
          "instructions"
        ],
        "title": "TelegramLink",
        "type": "object"
      },
...
      "UserSettings": {
        "properties": {
          "alerts": {
            "items": {
              "enum": [
                "brief",
                "injury",
                "waiver"
              ],
              "type": "string"
            },
            "title": "Alerts",
            "type": "array"
          },
          "awakeEnd": {
            "description": "HH:MM, 24-hour, in time_zone",
            "title": "Awakeend",
            "type": "string"
          },
          "awakeStart": {
            "description": "HH:MM, 24-hour, in time_zone",
            "title": "Awakestart",
            "type": "string"
          },
          "briefEnabled": {
            "title": "Briefenabled",
            "type": "boolean"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "draftStrategy": {
            "description": "all, or punt_<category>",
            "title": "Draftstrategy",
            "type": "string"
          },
          "telegramLinked": {
            "title": "Telegramlinked",
            "type": "boolean"
          },
          "theme": {
            "enum": [
              "system",
              "light",
              "dark"
            ],
            "title": "Theme",
            "type": "string"
          },
          "timeZone": {
            "title": "Timezone",
            "type": "string"
          },
          "version": {
            "description": "0 until first saved; the ETag carries it too",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "timeZone",
          "briefEnabled",
          "awakeStart",
          "awakeEnd",
          "alerts",
          "draftStrategy",
          "theme",
          "displayName",
          "telegramLinked",
          "version"
        ],
        "title": "UserSettings",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only views over the published daily outputs (D-62), plus users, roles and invites (D-64).",
    "title": "NBA Fantasy Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/invites": {
      "get": {
        "operationId": "list_invites_invites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Invite"
                  },
                  "title": "Response List Invites Invites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Invites",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_invite_invites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "An open invite for this email already existed"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Invite",
        "tags": [
          "users"
        ]
      }
    },
    "/invites/{invite_id}": {
      "delete": {
        "operationId": "revoke_invite_invites__invite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "title": "Invite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Invite",
        "tags": [
          "users"
        ]
      }
    },
...
    "/me": {
      "delete": {
        "operationId": "delete_me_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "The last owner can't delete their account"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "settings"
        ]
      },
      "get": {
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "users"
        ]
      }
    },
    "/me/export": {
      "get": {
        "operationId": "export_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountExport"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export",
        "tags": [
          "settings"
        ]
      }
    },
    "/me/settings": {
      "get": {
        "operationId": "get_settings_me_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "patch_settings_me_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "412": {
            "description": "Changed elsewhere: the body's `current` is the stored document"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "description": "If-Match is required"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/me/telegram/link": {
      "post": {
        "operationId": "telegram_link_me_telegram_link_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLink"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Telegram Link",
        "tags": [
          "settings"
        ]
      }
    },
    "/members": {
      "get": {
        "operationId": "list_members_members_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  },
                  "title": "Response List Members Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Members",
        "tags": [
          "users"
        ]
      }
    },
    "/members/{uid}": {
      "delete": {
        "operationId": "remove_member_members__uid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "uid",
            "required": true,
            "schema": {
              "title": "Uid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Member",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}